refactor: centralize RUN segment parsing
